requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.115.5
openai==1.93.0
pillow==11.2.1
pydantic==2.10.6
python-dotenv==1.1.0
requests==2.32.4
streamlit==1.45.1
uvicorn==0.34.3
